requirements.txt
pendulum==3.1.0
psycopg2==2.9.10
pymongo==4.15.1
python-dotenv==1.1.1
pandas==2.1.1